requirements.txt
ipython
num2words
omegaconf
pydub
PyYAML
edge-tts
fairseq==0.12.2
faiss_cpu==1.7.4
librosa==0.9.1
numpy==1.23.5
praat-parselmouth==0.4.3
pyworld==0.3.4
torchcrepe==0.0.20